# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: iHR Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /jobs:
    get:
      operationId: listJobs
      tags: [jobs]
      summary: List public job postings
      parameters:
        - name: companySlug
          in: query
          schema:
            type: ["string", "null"]
        - name: department
          in: query
          schema:
            type: ["string", "null"]
        - name: location
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Jobs list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Job"
    post:
      operationId: createJob
      tags: [jobs]
      summary: Create a new job posting
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
  /jobs/{id}:
    get:
      operationId: getJob
      tags: [jobs]
      summary: Get job by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
    patch:
      operationId: updateJob
      tags: [jobs]
      summary: Update job posting
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
    delete:
      operationId: deleteJob
      tags: [jobs]
      summary: Delete job posting
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /employees:
    get:
      operationId: listEmployees
      tags: [employees]
      summary: List employees
      parameters:
        - name: companyId
          in: query
          schema:
            type: ["integer", "null"]
        - name: department
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Employees list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Employee"
    post:
      operationId: createEmployee
      tags: [employees]
      summary: Create employee (onboarding)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmployeeInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
  /employees/{id}:
    get:
      operationId: getEmployee
      tags: [employees]
      summary: Get employee by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Employee
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
    patch:
      operationId: updateEmployee
      tags: [employees]
      summary: Update employee
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmployeeUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
...
  /leaves:
    get:
      operationId: listLeaves
      tags: [leaves]
      summary: List leave requests
      parameters:
        - name: employeeId
          in: query
          schema:
            type: ["integer", "null"]
        - name: status
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Leave requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Leave"
    post:
      operationId: createLeave
      tags: [leaves]
      summary: Apply for leave
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeaveInput"
      responses:
        "201":
          description: Leave applied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Leave"
  /leaves/{id}/status:
    patch:
      operationId: updateLeaveStatus
      tags: [leaves]
      summary: Approve or reject leave
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeaveStatusUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Leave"
...
components:
  schemas:
...
    Job:
      type: object
      required: [id, title, companyId, department, employmentType, status]
      properties:
        id:
          type: integer
        title:
          type: string
        companyId:
          type: integer
        companyName:
          type: ["string", "null"]
        department:
          type: string
        employmentType:
          type: string
          enum: [full_time, part_time, contract, internship]
        experienceLevel:
          type: ["string", "null"]
        salaryMin:
          type: ["integer", "null"]
        salaryMax:
          type: ["integer", "null"]
        skills:
          type: ["array", "null"]
          items:
            type: string
        description:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        workMode:
          type: ["string", "null"]
          enum: [remote, hybrid, onsite, null]
        openings:
          type: ["integer", "null"]
        deadline:
          type: ["string", "null"]
        status:
          type: string
          enum: [open, closed, draft]
        applicationsCount:
          type: ["integer", "null"]
        createdAt:
          type: string
    JobInput:
      type: object
      required: [title, companyId, department, employmentType, description]
      properties:
        title:
          type: string
        companyId:
          type: integer
        department:
          type: string
        employmentType:
          type: string
        experienceLevel:
          type: ["string", "null"]
        salaryMin:
          type: ["integer", "null"]
        salaryMax:
          type: ["integer", "null"]
        skills:
          type: ["array", "null"]
          items:
            type: string
        description:
          type: string
        location:
          type: ["string", "null"]
        workMode:
          type: ["string", "null"]
        openings:
          type: ["integer", "null"]
        deadline:
          type: ["string", "null"]
    JobUpdate:
      type: object
      properties:
        title:
          type: ["string", "null"]
        department:
          type: ["string", "null"]
        status:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        deadline:
          type: ["string", "null"]
...
    Employee:
      type: object
      required: [id, employeeCode, fullName, companyId, department, status]
      properties:
        id:
          type: integer
        employeeCode:
          type: string
        fullName:
          type: string
        firstName:
          type: ["string", "null"]
        middleName:
          type: ["string", "null"]
        lastName:
          type: ["string", "null"]
        email:
          type: string
        companyId:
          type: integer
        department:
          type: string
        designation:
          type: ["string", "null"]
        jobTitle:
          type: ["string", "null"]
        reportingManager:
          type: ["string", "null"]
        reportingManagerId:
          type: ["integer", "null"]
        salary:
          type: ["integer", "null"]
        basicSalary:
          type: ["integer", "null"]
        salaryUpdatedDate:
          type: ["string", "null"]
        joiningDate:
          type: ["string", "null"]
        positionUpdatedDate:
          type: ["string", "null"]
        dob:
          type: ["string", "null"]
        doe:
          type: ["string", "null"]
        ofcEmail:
          type: ["string", "null"]
        ofcNumber:
          type: ["string", "null"]
        status:
          type: string
          enum: [active, inactive, on_leave]
        avatarUrl:
          type: ["string", "null"]
        leaveBalance:
          type: ["integer", "null"]
    EmployeeInput:
      type: object
      required: [fullName, email, companyId, department]
      properties:
        fullName:
          type: string
        firstName:
          type: ["string", "null"]
        middleName:
          type: ["string", "null"]
        lastName:
          type: ["string", "null"]
        email:
          type: string
        companyId:
          type: integer
        department:
          type: string
        designation:
          type: ["string", "null"]
        jobTitle:
          type: ["string", "null"]
        salary:
          type: ["integer", "null"]
        basicSalary:
          type: ["integer", "null"]
        joiningDate:
          type: ["string", "null"]
        dob:
          type: ["string", "null"]
        doe:
          type: ["string", "null"]
        ofcEmail:
          type: ["string", "null"]
        ofcNumber:
          type: ["string", "null"]
        reportingManager:
          type: ["string", "null"]
        reportingManagerId:
          type: ["integer", "null"]
    EmployeeUpdate:
      type: object
      properties:
        designation:
          type: ["string", "null"]
        jobTitle:
          type: ["string", "null"]
        department:
          type: ["string", "null"]
        salary:
          type: ["integer", "null"]
        basicSalary:
          type: ["integer", "null"]
        salaryUpdatedDate:
          type: ["string", "null"]
        status:
          type: ["string", "null"]
        reportingManager:
          type: ["string", "null"]
        reportingManagerId:
          type: ["integer", "null"]
        joiningDate:
          type: ["string", "null"]
        positionUpdatedDate:
          type: ["string", "null"]
        dob:
          type: ["string", "null"]
        doe:
          type: ["string", "null"]
        ofcEmail:
          type: ["string", "null"]
        ofcNumber:
          type: ["string", "null"]
        firstName:
          type: ["string", "null"]
        middleName:
          type: ["string", "null"]
        lastName:
          type: ["string", "null"]
...
    Leave:
      type: object
      required: [id, employeeId, leaveType, startDate, endDate, status]
      properties:
        id:
          type: integer
        employeeId:
          type: integer
        employeeName:
          type: ["string", "null"]
        leaveType:
          type: string
          enum: [sick, casual, earned, maternity, paternity]
        startDate:
          type: string
        endDate:
          type: string
        days:
          type: ["integer", "null"]
        reason:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, approved, rejected]
        reviewedBy:
          type: ["string", "null"]
    LeaveInput:
      type: object
      required: [employeeId, leaveType, startDate, endDate]
      properties:
        employeeId:
          type: integer
        leaveType:
          type: string
        startDate:
          type: string
        endDate:
          type: string
        reason:
          type: ["string", "null"]
    LeaveStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [approved, rejected]
        reviewedBy:
          type: ["string", "null"]
...